+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - cli: Show all events in today by default
 - core: `EventConditions.startable` expect a `DateTimeAnchor` now
+- core: `TodoConditions.due` expect a `DateTimeAnchor` now
 
 ### Fixed